deleted leetcode.go move to index.go
